fix: move governance.ts out of SDK, address CI and review feedback
- Move governance.ts from solana/ts/sdk/ to solana/ts/scripts/ so it
  doesn't break the SDK build (it depends on @certusone/wormhole-sdk
  and local IDL artifacts not available in CI)
- Update import paths in all scripts that reference it
- Fix IDL import paths (../../target -> ../../../target)
- Add .env to evm/ts-scripts/.gitignore to prevent secret leakage
- Make guardianKey lazy (getGuardianKey function) so scripts that don't
  need it aren't forced to set the env var
- Add "private": true to evm/ts-scripts/package.json
- Fix prettier violation in evm env.ts provider construction
