+++ evm/ts-scripts/package.json
@@ -1,6 +1,7 @@
 {
   "name": "ts-scripts",
   "version": "1.0.0",
+  "private": true,
   "description": "",
   "main": "index.js",
   "scripts": {
