+++ solana/ts/scripts/env.ts
@@ -79,4 +79,6 @@ export function loadScriptConfig(filename: string): any {
   return config;
 }
 
-export const guardianKey = getEnv("GUARDIAN_KEY");
+export function getGuardianKey(): string {
+  return getEnv("GUARDIAN_KEY");
+}
