+++ solana/ts/scripts/executeGovernanceVaa.ts
@@ -11,7 +11,7 @@ import {
 } from "./env";
 import { addLedgerSignature, ledgerSignAndSend } from "./helpers";
 import { postVaaSolana } from "@certusone/wormhole-sdk";
-import { NTTGovernance } from "../sdk/governance";
+import { NTTGovernance } from "./governance";
 
 (async () => {
   const { vaa } = getGovernanceVaa();
